+++ src/main/java/com/marklogic/mule/extension/Operations.java
@@ -25,6 +25,9 @@ import com.marklogic.client.io.marker.JSONWriteHandle;
 import com.marklogic.client.query.DeleteQueryDefinition;
 import com.marklogic.client.query.QueryDefinition;
 import com.marklogic.client.query.QueryManager;
+import com.marklogic.mule.extension.api.DocumentAttributes;
+import com.marklogic.mule.extension.api.QueryFormat;
+import com.marklogic.mule.extension.api.QueryType;
 import org.mule.runtime.extension.api.annotation.param.Connection;
 import org.mule.runtime.extension.api.annotation.param.Content;
 import org.mule.runtime.extension.api.annotation.param.MediaType;
