+++ src/main/java/com/marklogic/mule/extension/api/ConnectionType.java
@@ -13,9 +13,23 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.marklogic.mule.extension;
+package com.marklogic.mule.extension.api;
 
-public enum QueryFormat {
-    JSON,
-    XML;
+import com.marklogic.client.DatabaseClient;
+
+public enum ConnectionType {
+    DIRECT {
+        @Override
+        public DatabaseClient.ConnectionType getMarkLogicConnectionType() {
+            return DatabaseClient.ConnectionType.DIRECT;
+        }
+    },
+    GATEWAY {
+        @Override
+        public DatabaseClient.ConnectionType getMarkLogicConnectionType() {
+            return DatabaseClient.ConnectionType.GATEWAY;
+        }
+    };
+
+    public abstract DatabaseClient.ConnectionType getMarkLogicConnectionType();
 }
