+++ src/test/java/com/marklogic/mule/extension/AbstractFlowTester.java
@@ -1,5 +1,6 @@
 package com.marklogic.mule.extension;
 
+import com.marklogic.mule.extension.api.DocumentAttributes;
 import org.mule.functional.junit4.MuleArtifactFunctionalTestCase;
 import org.mule.runtime.api.message.Message;
