+++ src/test/java/com/marklogic/mule/extension/MetadataVerifier.java
@@ -1,5 +1,6 @@
 package com.marklogic.mule.extension;
 
+import com.marklogic.mule.extension.api.DocumentAttributes;
 import org.junit.Assert;
 
 import javax.xml.namespace.QName;
