+++ src/test/java/com/marklogic/mule/extension/WriteDocumentTest.java
@@ -1,6 +1,7 @@
 package com.marklogic.mule.extension;
 
 
+import com.marklogic.mule.extension.api.DocumentAttributes;
 import org.junit.Test;
 
 import java.util.List;
